+++ src/audio/CMakeLists.txt
@@ -177,6 +177,7 @@ set(sof_audio_modules mixer volume src asrc eq-fir eq-iir dcblock crossover tdfb
 # sources for each module
 if(CONFIG_IPC_MAJOR_3)
 	set(volume_sources volume/volume.c volume/volume_generic.c volume/volume_ipc3.c)
+	set(asrc_sources asrc/asrc_ipc3.c)
 	set(src_sources src/src.c src/src_ipc3.c src/src_generic.c)
 	set(eq-iir_sources eq_iir/eq_iir_ipc3.c eq_iir/eq_iir_generic.c)
 	set(eq-fir_sources eq_fir/eq_fir_ipc3.c)
@@ -185,6 +186,7 @@ if(CONFIG_IPC_MAJOR_3)
 	set(dcblock_sources dcblock/dcblock_ipc3.c)
 elseif(CONFIG_IPC_MAJOR_4)
 	set(volume_sources volume/volume.c volume/volume_generic.c volume/volume_ipc4.c)
+	set(asrc_sources asrc/asrc_ipc4.c)
 	set(src_sources src/src.c src/src_ipc4.c src/src_generic.c)
 	set(eq-iir_sources eq_iir/eq_iir_ipc4.c eq_iir/eq_iir_generic.c)
 	set(eq-fir_sources eq_fir/eq_fir_ipc4.c)
