+++ src/audio/asrc/CMakeLists.txt
@@ -3,3 +3,8 @@
 add_local_sources(sof asrc.c asrc_farrow.c asrc_farrow_generic.c
 	asrc_farrow_hifi3.c)
 
+if(CONFIG_IPC_MAJOR_3)
+	add_local_sources(sof asrc_ipc3.c)
+elseif(CONFIG_IPC_MAJOR_4)
+	add_local_sources(sof asrc_ipc4.c)
+endif()
