+++ zephyr/CMakeLists.txt
@@ -510,6 +510,14 @@ zephyr_library_sources_ifdef(CONFIG_COMP_ASRC
 	${SOF_AUDIO_PATH}/asrc/asrc_farrow_generic.c
 )
 
+if(CONFIG_COMP_ASRC)
+	if(CONFIG_IPC_MAJOR_3)
+		zephyr_library_sources(${SOF_AUDIO_PATH}/asrc/asrc_ipc3.c)
+	elseif(CONFIG_IPC_MAJOR_4)
+		zephyr_library_sources(${SOF_AUDIO_PATH}/asrc/asrc_ipc4.c)
+	endif()
+endif()
+
 zephyr_library_sources_ifdef(CONFIG_COMP_DCBLOCK
 	${SOF_AUDIO_PATH}/dcblock/dcblock_generic.c
 	${SOF_AUDIO_PATH}/dcblock/dcblock.c
